refactor: move statics to Class40_Sub3
